Uppdatera delete-funktionen

+++ functions/deleteBooking/index.js
@@ -4,25 +4,34 @@ import { db } from '../../services/db.js';
 export async function handler(event, context) {
     try {
 
-        // Control if the booking id is missing in the request
+        // Kontrollera att boknings-ID:t finns
         if(!event.pathParameters || !event.pathParameters.id) {
             return sendError(400,  { success: false, message: 'Missing booking id' });
         }
 
-        // Remove the booking from the database
-       const result = await db.delete({
+        const bookingId = event.pathParameters.id;
+
+        const getBooking = await db.get({
+            TableName: 'bookings-db',
+            Key: {
+                id: bookingId
+            }
+        }).promise();
+
+        // Kontrollera att bokningen finns
+        if(!getBooking.Item) {
+            return sendError(404, { success: false, message: 'Booking not found' });
+        }
+
+        // Ta bort bokningen från databasen
+       await db.delete({
         TableName: 'bookings-db',
         Key: {
-            id: event.pathParameters.id
+            id: bookingId
         }
        }).promise();
-
-       // If the booking is not found, return an error
-       if(!result.Attributes === undefined) {
-              return sendError(404, { success: false, message: 'Booking not found' });
-       }
-        
-       return sendResponse(200, { success: true, message: 'Booking successfully removed' });
+       
+       return sendResponse(200, { success: true, message: 'Booking successfully deleted' });
        
         } catch (error) {
             return sendError(500, { success: false, message: 'Could not delete booking' });
